Use unsigned integers for vector sizes and capacities (#203)
* Use unsigned integers for vector sizes and capacities

* Cast to unsigned int for malloc argument to avoid compiler warnings

* Change size types to unsigned, add casts where required and address other compiler warnings.

* Avoid problems with some compilers

* Try to make windows happy

* More signed/unsigned fixes

* Add generated lexer and parser

* Use _strdup on windows

* More casts

* Remove unimplemented member functions

* Formatting

* Enable "narrowing-conversions" clang-tidy changes and resolve warnings

* Update generated parser

* Fix clang-tidy problems in the parser

---------

+++ chuffed/support/heap.h
@@ -22,6 +22,8 @@ OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWA
 
 #include "chuffed/support/vec.h"
 
+#include <cassert>
+
 //=================================================================================================
 // A heap implementation with support for decrease/increase key.
 
@@ -49,9 +51,10 @@ class Heap {
 
 	inline void percolateDown(int i) {
 		const int x = heap[i];
-		while (left(i) < heap.size()) {
+		while (left(i) < static_cast<int>(heap.size())) {
 			const int child =
-					right(i) < heap.size() && lt(heap[right(i)], heap[left(i)]) ? right(i) : left(i);
+					right(i) < static_cast<int>(heap.size()) && lt(heap[right(i)], heap[left(i)]) ? right(i)
+																																												: left(i);
 			if (!lt(heap[child], x)) {
 				break;
 			}
@@ -73,9 +76,9 @@ public:
 
 	int size() const { return heap.size(); }
 	bool empty() const { return heap.size() == 0; }
-	bool inHeap(int n) const { return n < indices.size() && indices[n] >= 0; }
+	bool inHeap(int n) const { return n < static_cast<int>(indices.size()) && indices[n] >= 0; }
 	int operator[](int index) const {
-		assert(index < heap.size());
+		assert(index < static_cast<int>(heap.size()));
 		return heap[index];
 	}
